Refactor: update image name

+++ training/build_push.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
-docker build -t asdfry/dali-example:imagenet .
-yes | docker image prune
-docker images
-docker push asdfry/dali-example:imagenet
+if [ $# = 0 ]; then
+    echo "Date argument required"
+elif [ $# = 1 ]; then
+    docker build -t asdfry/train-resnet:$1 .
+    yes | docker image prune
+    docker images
+    docker push asdfry/train-resnet:$1
+fi
